+++ lib/typed_soup.dart
@@ -1,5 +1,6 @@
 library;
 
-export 'package:typed_soup/src/ts_element.dart';
-export 'package:typed_soup/src/ts_soup.dart';
-export 'package:typed_soup/src/extensions.dart';
+export 'src/ts_soup.dart';
+export 'src/ts_element.dart';
+export 'src/soup_strainer.dart';
+export 'src/extensions.dart';